Move LowStockProduct into separate component
