+++ dashboard/src/components/overview/LowStockProduct.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import { Pressable, StyleSheet } from 'react-native';
+import Typography from '../global/Typography';
+import { formatNaira } from '../../utils/currency';
+import CustomImage from '../global/CustomImage';
+import { ProductsQuery } from '../../types/api';
+
+interface LowStockProductProps {
+	onPress(): void;
+	product: ProductsQuery['currentStore']['products'][number];
+}
+
+const LowStockProduct: React.FC<LowStockProductProps> = ({
+	onPress,
+	product
+}) => {
+	return (
+		<Pressable onPress={onPress} style={{ marginRight: 8, width: 160 }}>
+			<CustomImage uri={product.images[0]?.path} size={160} />
+			<Typography weight='medium' style={styles.text} ellipsize>
+				{product.name}
+			</Typography>
+			<Typography variant='label' style={styles.text} ellipsize>
+				{formatNaira(product.unitPrice)}
+			</Typography>
+		</Pressable>
+	);
+};
+
+const styles = StyleSheet.create({
+	text: {
+		marginTop: 2
+	}
+});
+
+export default LowStockProduct;
